requirements.txt
absl-py==2.3.1
action-msgs==2.0.3
action-tutorials-interfaces==0.33.6
action-tutorials-py==0.33.6
actionlib-msgs==5.3.6
actuator-msgs==0.0.1
addict==2.4.0
aiofiles==24.1.0
ament-cmake-test==2.5.4
ament-copyright==0.17.3
ament-cppcheck==0.17.3
ament-cpplint==0.17.3
ament-flake8==0.17.3
ament-index-python==1.8.1
ament-lint==0.17.3
ament-lint-cmake==0.17.3
ament-package==0.16.4
ament-pep257==0.17.3
ament-uncrustify==0.17.3
ament-xmllint==0.17.3
angles==1.16.0
annotated-types==0.7.0
anyio==4.11.0
basicsr==1.4.2
Brotli==1.1.0
builtin-interfaces==2.0.3
camera-calibration==5.0.11
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
composition-interfaces==2.0.3
contourpy==1.3.2
cv-bridge==4.1.0
cycler==0.12.1
demo-nodes-py==0.33.6
diagnostic-msgs==5.3.6
diagnostic-updater==4.2.6
domain-coordinator==0.12.0
example-interfaces==0.12.0
examples-rclpy-executors==0.19.6
examples-rclpy-minimal-action-client==0.19.6
examples-rclpy-minimal-action-server==0.19.6
examples-rclpy-minimal-client==0.19.6
examples-rclpy-minimal-publisher==0.19.6
examples-rclpy-minimal-service==0.19.6
examples-rclpy-minimal-subscriber==0.19.6
exceptiongroup==1.3.0
facexlib==0.3.0
fastapi==0.117.1
ffmpy==0.6.1
filelock==3.19.1
filterpy==1.4.5
fonttools==4.60.0
fsspec==2025.9.0
future==1.0.0
geometry-msgs==5.3.6
gfpgan==1.3.8
gps-msgs==2.1.1
gradio==5.47.0
gradio_client==1.13.2
gradio_imageslider==0.0.20
groovy==0.1.2
grpcio==1.75.0
h11==0.16.0
hf-xet==1.1.10
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.35.1
idna==3.10
image-geometry==4.1.0
imageio==2.37.0
interactive-markers==2.5.4
Jinja2==3.1.6
kiwisolver==1.4.9
laser-geometry==2.7.1
launch==3.4.6
launch-ros==0.26.9
launch-testing==3.4.6
launch-testing-ros==0.26.9
launch-xml==3.4.6
launch-yaml==3.4.6
lazy_loader==0.4
lifecycle-msgs==2.0.3
llvmlite==0.45.0
lmdb==1.7.3
logging-demo==0.33.6
map-msgs==2.4.1
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.6
mdurl==0.1.2
message-filters==4.11.7
mpmath==1.3.0
nav-msgs==5.3.6
networkx==3.4.2
numba==0.62.0
numpy==1.26.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
opencv-python==4.11.0.86
orjson==3.11.3
osrf-pycommon==2.1.7
packaging==25.0
pandas==2.3.2
pcl-msgs==1.0.0
pendulum-msgs==0.33.6
pillow==11.3.0
platformdirs==4.4.0
protobuf==6.32.1
pydantic==2.11.9
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-multipart==0.0.20
python-qt-binding==2.2.1
pytz==2025.2
PyYAML==6.0.2
qt-dotgraph==2.7.5
qt-gui==2.7.5
qt-gui-cpp==2.7.5
qt-gui-py-common==2.7.5
quality-of-service-demo-py==0.33.6
rcl-interfaces==2.0.3
rclpy==7.1.5
rcutils==6.7.4
realesrgan==0.3.0
requests==2.32.5
resource-retriever==3.4.4
rich==14.1.0
rmw-dds-common==3.1.0
ros-gz-bridge==1.0.16
ros-gz-interfaces==1.0.16
ros-gz-sim==1.0.16
ros2action==0.32.6
ros2bag==0.26.9
ros2bag-mcap-cli==0.26.9
ros2bag-sqlite3-cli==0.26.9
ros2cli==0.32.6
ros2component==0.32.6
ros2doctor==0.32.6
ros2interface==0.32.6
ros2launch==0.26.9
ros2lifecycle==0.32.6
ros2multicast==0.32.6
ros2node==0.32.6
ros2param==0.32.6
ros2pkg==0.32.6
ros2run==0.32.6
ros2service==0.32.6
ros2topic==0.32.6
rosbag2-interfaces==0.26.9
rosbag2-py==0.26.9
rosgraph-msgs==2.0.3
rosidl-adapter==4.6.6
rosidl-cli==4.6.6
rosidl-cmake==4.6.6
rosidl-generator-c==4.6.6
rosidl-generator-cpp==4.6.6
rosidl-generator-py==0.22.1
rosidl-generator-type-description==4.6.6
rosidl-parser==4.6.6
rosidl-pycommon==4.6.6
rosidl-runtime-py==0.13.1
rosidl-typesupport-c==3.2.2
rosidl-typesupport-cpp==3.2.2
rosidl-typesupport-fastrtps-c==3.6.2
rosidl-typesupport-fastrtps-cpp==3.6.2
rosidl-typesupport-introspection-c==4.6.6
rosidl-typesupport-introspection-cpp==4.6.6
rpyutils==0.4.1
rqt==1.6.0
rqt-action==2.2.0
rqt-bag==1.5.5
rqt-bag-plugins==1.5.5
rqt-console==2.2.1
rqt-graph==1.5.5
rqt-gui==1.6.0
rqt-gui-py==1.6.0
rqt-msg==1.5.1
rqt-plot==1.4.4
rqt-publisher==1.7.2
rqt-py-common==1.6.0
rqt-py-console==1.2.2
rqt-reconfigure==1.6.2
rqt-service-caller==1.2.1
rqt-shell==1.2.2
rqt-srv==1.2.2
rqt-topic==1.7.3
ruff==0.13.1
safehttpx==0.1.6
scikit-image==0.25.2
scipy==1.15.3
semantic-version==2.10.0
sensor-msgs==5.3.6
sensor-msgs-py==5.3.6
service-msgs==2.0.3
shape-msgs==5.3.6
shellingham==1.5.4
sick-safetyscanners2-interfaces==1.0.0
six==1.17.0
sniffio==1.3.1
sros2==0.13.4
starlette==0.48.0
statistics-msgs==2.0.3
std-msgs==5.3.6
std-srvs==5.3.6
stereo-msgs==5.3.6
sympy==1.14.0
tb-nightly==2.21.0a20250925
teleop-twist-keyboard==2.4.0
tensorboard-data-server==0.7.2
tf-transformations==1.1.0
tf2-geometry-msgs==0.36.14
tf2-kdl==0.36.14
tf2-msgs==0.36.14
tf2-py==0.36.14
tf2-ros-py==0.36.14
tf2-sensor-msgs==0.36.14
tf2-tools==0.36.14
theora-image-transport==4.0.6
tifffile==2025.5.10
tomli==2.2.1
tomlkit==0.13.3
topic-monitor==0.33.6
torch==1.13.1
torchvision==0.14.1
tqdm==4.67.1
trajectory-msgs==5.3.6
triton==3.4.0
turtlesim==1.8.3
type-description-interfaces==2.0.3
typer==0.19.2
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
unique-identifier-msgs==2.5.0
urllib3==2.5.0
uvicorn==0.37.0
vision-msgs==4.1.1
visualization-msgs==5.3.6
websockets==15.0.1
Werkzeug==3.1.3
xacro==2.1.1
yapf==0.43.0
